Move PrettyPrinter inside of JsonDiff
